requirements.txt
rasa==3.6.21
rasa-sdk==3.6.2
tensorflow==2.12.0
python-dateutil==2.8.2
requests==2.32.4
psycopg2-binary==2.9.10
pypdf==5.1.0
reportlab
pandas
openpyxl
pillow
python-docx
websocket-client